isArmnummerAndStrassenseiteInvalid moved
